Fix builtin macros split exprs on comma
